requirements.txt
# Warp-Claw Dependencies

# Core
torch>=2.0.0
transformers>=4.35.0
huggingface-hub>=0.20.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Tools
aiohttp>=3.9.0
aiofiles>=23.2.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0

# Optional: Vector database
# lancedb>=0.8.0
# chromadb>=0.4.0

# Dashboard (optional)
streamlit>=1.30.0

# WebSocket (optional)
websockets>=12.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.26.0